requirements.txt
fastapi
uvicorn
pillow
numpy
pandas
tflite-runtime
python-multipart
